+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 - Added SDL2 wrapper (``core/sdl_wrapper.h``, ``core/sdl_wrapper.cpp``)
 - Removed graphics (SDL2 interaction) wrapper (``core/graphics.h``, ``core/graphics.cpp``)
 - Added class for easy drawing 2D image via buffer and OpenGL (``core/gl_image.h``, ``core/gl_image.cpp``)
+- Added RGB OpenGL triangle (``core/rgb_triangle.h``, ``core/rgb_triangle.cpp``)
+- Added keyboard interaction for RGB triangle
 
 2.0.1 - 2022-11-22
 ------------------